Sales dashboard
A sales dashboard  provides a concise and visually appealing overview of key sales performance metrics and data. It is typically used by businesses and sales teams to monitor and analyze their sales activities and results. Here's a typical structure and the kind of information you might find in a sales dashboard